requirements.txt
streamlit>=1.28.0
playwright>=1.40.0
selenium>=4.15.0
paddleocr>=2.7.0
easyocr>=1.7.0
opencv-python>=4.8.0
pillow>=10.1.0
anthropic>=0.7.0
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.25.0
numpy>=1.24.0
aiohttp>=3.9.0
tqdm>=4.66.0
